+++ qc-application/src/main/java/com/eager/questioncloud/application/api/hub/QuestionHubReviewController.java
@@ -7,9 +7,9 @@ import com.eager.questioncloud.application.api.hub.QuestionHubReviewControllerRe
 import com.eager.questioncloud.application.api.hub.QuestionHubReviewControllerResponse.MyQuestionReviewResponse;
 import com.eager.questioncloud.application.security.UserPrincipal;
 import com.eager.questioncloud.core.common.PagingInformation;
-import com.eager.questioncloud.core.domain.review.MyQuestionReview;
-import com.eager.questioncloud.core.domain.review.QuestionReview;
-import com.eager.questioncloud.core.domain.review.QuestionReviewDetail;
+import com.eager.questioncloud.core.domain.review.dto.MyQuestionReview;
+import com.eager.questioncloud.core.domain.review.dto.QuestionReviewDetail;
+import com.eager.questioncloud.core.domain.review.model.QuestionReview;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
