+++ qc-application/src/main/java/com/eager/questioncloud/application/api/hub/QuestionHubReviewRegister.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.api.hub;
 
 import com.eager.questioncloud.core.domain.question.infrastructure.QuestionRepository;
-import com.eager.questioncloud.core.domain.review.QuestionReview;
-import com.eager.questioncloud.core.domain.review.QuestionReviewRepository;
+import com.eager.questioncloud.core.domain.review.infrastructure.QuestionReviewRepository;
+import com.eager.questioncloud.core.domain.review.model.QuestionReview;
 import com.eager.questioncloud.core.domain.userquestion.UserQuestionRepository;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
