+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/review/implement/QuestionReviewStatisticsGenerator.java
@@ -1,6 +1,8 @@
-package com.eager.questioncloud.core.domain.review;
+package com.eager.questioncloud.core.domain.review.implement;
 
 import com.eager.questioncloud.core.domain.question.event.RegisteredQuestionEvent;
+import com.eager.questioncloud.core.domain.review.infrastructure.QuestionReviewStatisticsRepository;
+import com.eager.questioncloud.core.domain.review.model.QuestionReviewStatistics;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
