+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/review/implement/QuestionReviewStatisticsUpdater.java
@@ -1,5 +1,10 @@
-package com.eager.questioncloud.core.domain.review;
+package com.eager.questioncloud.core.domain.review.implement;
 
+import com.eager.questioncloud.core.domain.review.event.DeletedReviewEvent;
+import com.eager.questioncloud.core.domain.review.event.ModifiedReviewEvent;
+import com.eager.questioncloud.core.domain.review.event.RegisteredReviewEvent;
+import com.eager.questioncloud.core.domain.review.infrastructure.QuestionReviewStatisticsRepository;
+import com.eager.questioncloud.core.domain.review.model.QuestionReviewStatistics;
 import com.eager.questioncloud.lock.LockKeyGenerator;
 import com.eager.questioncloud.lock.LockManager;
 import lombok.RequiredArgsConstructor;
