+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/review/infrastructure/QuestionReviewStatisticsJpaRepository.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.review;
+package com.eager.questioncloud.core.domain.review.infrastructure;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
