+++ qc-core/src/test/java/com/eager/questioncloud/core/domain/creator/CreatorStatisticsProcessorTest.java
@@ -10,9 +10,9 @@ import com.eager.questioncloud.core.domain.creator.model.CreatorStatistics;
 import com.eager.questioncloud.core.domain.question.QuestionBuilder;
 import com.eager.questioncloud.core.domain.question.infrastructure.QuestionRepository;
 import com.eager.questioncloud.core.domain.question.model.Question;
-import com.eager.questioncloud.core.domain.review.DeletedReviewEvent;
-import com.eager.questioncloud.core.domain.review.ModifiedReviewEvent;
-import com.eager.questioncloud.core.domain.review.RegisteredReviewEvent;
+import com.eager.questioncloud.core.domain.review.event.DeletedReviewEvent;
+import com.eager.questioncloud.core.domain.review.event.ModifiedReviewEvent;
+import com.eager.questioncloud.core.domain.review.event.RegisteredReviewEvent;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
